backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==1.50.0
python-dotenv==1.0.1
pydantic==2.9.0
gTTS==2.5.3
